requirements.txt
alembic==1.11.1
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
hiredis==2.2.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
llvmlite==0.40.0
Mako==1.2.4
MarkupSafe==2.1.2
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyarrow==12.0.0
PyJWT==2.7.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.15
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
xgboost==1.7.5
zipp==3.15.0
